Method debug

+++ Editor/Method.xaml.cs
@@ -75,11 +75,15 @@ namespace Editor
 
         private void DeleteMethod_Click(object sender, RoutedEventArgs e)
         {
-            if (MessageBox.Show("Biztosan törlöd a(z) " + methodList.SelectedItem.ToString() + " eljárást?", "Eljárás törlés", MessageBoxButton.YesNo, MessageBoxImage.Asterisk) == MessageBoxResult.Yes)
+            if (methodList.SelectedItem != null)
             {
-                SQLiteHelper.GetSqlHelper().DeleteMethod(methodList.SelectedItem.ToString());
-                SetMethodNames();
-            } 
+                if (MessageBox.Show("Biztosan törlöd a(z) " + methodList.SelectedItem.ToString() + " eljárást?", "Eljárás törlés", MessageBoxButton.YesNo, MessageBoxImage.Asterisk) == MessageBoxResult.Yes)
+                {
+                    SQLiteHelper.GetSqlHelper().DeleteMethod(methodList.SelectedItem.ToString());
+                    SetMethodNames();
+                }
+            }
+            prevItem = null;
         }
 
         private bool TextChanged()
